[FF] website_sale_uom_dimension

+++ website_sale_uom_dimension/controllers/main.py
@@ -1,54 +1,24 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
 import json
-import logging
-from werkzeug.exceptions import NotFound
 
-from odoo import fields, http
+from odoo import http
 from odoo.http import request
-from odoo.addons.website.controllers.main import QueryURL
 from odoo.addons.website_sale.controllers.main import WebsiteSale
-
-_logger = logging.getLogger(__name__)
-
-PPG = 20  # Products Per Page
-PPR = 4   # Products Per Row
 
 
 class WebsiteSaleDimensions(WebsiteSale):
 
-    @http.route(['/shop/cart/update'], type='http', auth="public", methods=['POST'], website=True, csrf=False)
+    @http.route(['/shop/cart/update'], type='http', auth="public", methods=['POST'], website=True)
     def cart_update(self, product_id, add_qty=1, set_qty=0, **kw):
-        """This route is called when adding a product to cart (no options)."""
+        res = super().cart_update(product_id, add_qty, set_qty, **kw)
         sale_order = request.website.sale_get_order(force_create=True)
-        if sale_order.state != 'draft':
-            request.session['sale_order_id'] = None
-            sale_order = request.website.sale_get_order(force_create=True)
-
-        product_custom_attribute_values = None
-        if kw.get('product_custom_attribute_values'):
-            product_custom_attribute_values = json.loads(kw.get('product_custom_attribute_values'))
-
-        no_variant_attribute_values = None
-        if kw.get('no_variant_attribute_values'):
-            no_variant_attribute_values = json.loads(kw.get('no_variant_attribute_values'))
-
-        sale_order._cart_update(
-            product_id=int(product_id),
-            add_qty=add_qty,
-            set_qty=set_qty,
-            product_custom_attribute_values=product_custom_attribute_values,
-            no_variant_attribute_values=no_variant_attribute_values
-        )
         order_line = sale_order._cart_find_product_line(int(product_id), line_id = False, **kw)[:1]
-        if order_line and order_line.product_uom.dimension_ids:
+        if order_line and order_line.product_uom.dimension_ids and kw.get('dimension_values'):
+            dimension_values = json.loads(kw['dimension_values'])
             order_line.dimension_ids.unlink()
             vals = {
-                'dimension_ids': [(0, 0, {'dimension_id': d.id, 'quantity': kw.get('dimension-' + str(d.id), 0)}) for d in
-                                  order_line.product_uom.dimension_ids],
+                'dimension_ids': [(0, 0, {'dimension_id': d['dimension_id'], 'quantity': float(d['quantity'])}) for d in dimension_values],
                 'product_dimension_qty': add_qty
             }
             order_line.write(vals)
             order_line.onchange_dimension_ids()
-        return request.redirect("/shop/cart")
-
+        return res
